@@ -11,13 +11,13 @@
   },
   "dependencies": {
     "jsdom": "^26.1.0",
-    "vue": "^2.7",
-    "vuex": "^3.4.0"
+    "vue": "^3.5",
+    "vuex": "^4.1.0"
   },
   "devDependencies": {
     "@tailwindcss/vite": "^4.1.11",
-    "@vitejs/plugin-vue2": "^2.2.3",
-    "@vue/test-utils": "^1.3.6",
+    "@vitejs/plugin-vue": "^6.0.0",
+    "@vue/test-utils": "^2.4.0",
     "eslint": "^9.32.0",
     "eslint-plugin-vue": "^10.3.0",
     "vite": "^6.0.0",
