@@ -1,11 +1,6 @@
-import Vue from 'vue'
-import App from './App.vue'
-import './assets/tailwind.css'
-import store from './store'
+import { createApp } from "vue";
+import App from "./App.vue";
+import "./assets/tailwind.css";
+import store from "./store";
 
-Vue.config.productionTip = false
-
-new Vue({
-  store,
-  render: h => h(App)
-}).$mount('#app')
+createApp(App).mount("#app").use(store);
